backend/requirements.txt
flask
flask-sqlalchemy
flask_migrate
python-dotenv
psycopg2-binary
flask-bcrypt
PyJWT
requests
flask-cors